Move AddToDashSelectDashModal into common
