+++ frontend/src/metabase/common/components/Pickers/index.ts
@@ -1,4 +1,5 @@
 export * from "./EntityPicker";
+export { AddToDashSelectDashModal } from "./AddToDashSelectDashModal";
 export { CollectionPickerModal } from "./CollectionPicker";
 export { DashboardPickerModal } from "./DashboardPicker";
 export { DataPickerModal } from "./DataPicker";
