+++ frontend/src/metabase/query_builder/components/QueryModals/QueryModals.tsx
@@ -6,6 +6,7 @@ import _ from "underscore";
 import { getDashboard } from "metabase/api";
 import { useGetDefaultCollectionId } from "metabase/collections/hooks";
 import { Modal } from "metabase/common/components/Modal";
+import { AddToDashSelectDashModal } from "metabase/common/components/Pickers/AddToDashSelectDashModal";
 import { SaveQuestionModal } from "metabase/common/components/SaveQuestionModal";
 import { type ToastArgs, useToast } from "metabase/common/hooks";
 import { QuestionEmbedWidget } from "metabase/embedding/components/QuestionEmbedWidget";
@@ -28,8 +29,6 @@ import { Text } from "metabase/ui";
 import * as Urls from "metabase/urls";
 import Question from "metabase-lib/v1/Question";
 import type { Card, DashboardTabId } from "metabase-types/api";
-
-import { AddToDashSelectDashModal } from "../AddToDashSelectDashModal";
 
 type OnCreateOptions = { dashboardTabId?: DashboardTabId | undefined };
